MB-10533 Move cert set up/removal funcs to new auth file
This way we can have all auth-related code in a more central file. Will move some version of `LoginTaskSet` in here too.
