+++ locustfiles/prime.py
@@ -6,7 +6,8 @@ from locust.env import Environment
 from tasks import PrimeTasks, SupportTasks
 from utils.base import ImplementationError
 from utils.constants import INTERNAL_API_KEY, PRIME_API_KEY
-from utils.hosts import MilMoveDomain, MilMoveHostMixin, remove_certs, set_up_certs
+from utils.hosts import MilMoveDomain, MilMoveHostMixin
+from utils.auth import remove_certs, set_up_certs
 from utils.parsers import InternalAPIParser, PrimeAPIParser, SupportAPIParser
 from utils.types import LOCUST_RUNNER_TYPE
