+++ utils/tests/test_hosts.py
@@ -2,15 +2,12 @@
 """ Tests utils/hosts.py """
 import os
 import re
-from pathlib import Path
 from unittest import mock
 
 import pytest
-from _pytest.monkeypatch import MonkeyPatch
 
 from utils.base import ImplementationError
-from utils.constants import DP3_CERT_KEY_PEM_FILENAME
-from utils.hosts import MilMoveDomain, MilMoveEnv, MilMoveHostMixin, remove_certs, set_up_certs
+from utils.hosts import MilMoveDomain, MilMoveEnv, MilMoveHostMixin
 
 
 class TestMilMoveDomain:
@@ -106,93 +103,3 @@ class TestMilMoveHostMixin:
         assert re.search("tls/devlocal-mtls.key", self.TestUser1.cert_kwargs["cert"][1])
         assert re.search("tls/devlocal-mtls.cer", self.TestUser2.cert_kwargs["cert"][0])
         assert re.search("tls/devlocal-mtls.key", self.TestUser2.cert_kwargs["cert"][1])
-
-
-class TestSetUpCerts:
-    """
-    Tests for set_up_certs
-    """
-
-    def test_no_file_is_created_if_running_locally(self, tmp_path: Path) -> None:
-        cert_key_pem = tmp_path / DP3_CERT_KEY_PEM_FILENAME
-
-        with mock.patch("utils.hosts.DP3_CERT_KEY_PEM", cert_key_pem):
-            set_up_certs(host=MilMoveEnv.LOCAL.value)
-
-            assert len(list(tmp_path.iterdir())) == 0
-
-    @pytest.mark.parametrize("tls_env_var", ["MOVE_MIL_DP3_TLS_CERT", "MOVE_MIL_DP3_TLS_KEY"])
-    def test_missing_env_tls_var_raises_exception_for_deployed_host(
-        self, tls_env_var: str, tmp_path: Path, monkeypatch: MonkeyPatch
-    ) -> None:
-        monkeypatch.delenv(tls_env_var, raising=False)
-
-        cert_key_pem = tmp_path / DP3_CERT_KEY_PEM_FILENAME
-
-        with mock.patch("utils.hosts.DP3_CERT_KEY_PEM", cert_key_pem):
-            with pytest.raises(
-                ImplementationError,
-                match="Cannot run load testing in a deployed environment without the matching certificate and key.",
-            ):
-
-                set_up_certs(host=MilMoveEnv.DP3.value)
-
-            assert len(list(tmp_path.iterdir())) == 0
-
-    def test_cert_created_for_deployed_host_with_env_vars(self, tmp_path: Path, monkeypatch: MonkeyPatch):
-        fake_cert = "fake cert"
-        fake_key = "fake key"
-        monkeypatch.setenv("MOVE_MIL_DP3_TLS_CERT", fake_cert)
-        monkeypatch.setenv("MOVE_MIL_DP3_TLS_KEY", fake_key)
-
-        cert_key_pem = tmp_path / DP3_CERT_KEY_PEM_FILENAME
-
-        with mock.patch("utils.hosts.DP3_CERT_KEY_PEM", cert_key_pem):
-            set_up_certs(host=MilMoveEnv.DP3.value)
-
-            assert len(list(tmp_path.iterdir())) == 1
-
-            assert cert_key_pem.exists()
-
-            assert f"{fake_cert}\n{fake_key}" == cert_key_pem.read_text()
-
-
-class TestRemoveCerts:
-    """
-    Tests for remove_certs
-    """
-
-    def test_no_file_is_removed_if_running_locally(self, tmp_path: Path) -> None:
-        cert_key_pem = tmp_path / DP3_CERT_KEY_PEM_FILENAME
-
-        cert_key_pem.touch()  # creating it just so that we can test that it won't get removed
-
-        with mock.patch("utils.hosts.DP3_CERT_KEY_PEM", cert_key_pem):
-            remove_certs(host=MilMoveEnv.LOCAL.value)
-
-            assert len(list(tmp_path.iterdir())) == 1
-            assert cert_key_pem.exists()
-
-    def test_cert_removed_for_deployed_envs(self, tmp_path: Path):
-        cert_key_pem = tmp_path / DP3_CERT_KEY_PEM_FILENAME
-
-        cert_key_pem.touch()
-
-        with mock.patch("utils.hosts.DP3_CERT_KEY_PEM", cert_key_pem):
-            remove_certs(host=MilMoveEnv.DP3.value)
-
-            assert len(list(tmp_path.iterdir())) == 0
-
-            assert not cert_key_pem.exists()
-
-    def test_no_error_if_cert_already_gone(self, tmp_path: Path):
-        cert_key_pem = tmp_path / DP3_CERT_KEY_PEM_FILENAME
-
-        assert len(list(tmp_path.iterdir())) == 0
-
-        with mock.patch("utils.hosts.DP3_CERT_KEY_PEM", cert_key_pem):
-            remove_certs(host=MilMoveEnv.DP3.value)
-
-            assert len(list(tmp_path.iterdir())) == 0
-
-            assert not cert_key_pem.exists()
